-refactor create teams component to teams component (#64)
- refactor functions that calculate players and team scores on the FE to utils
- add tests for team stats utils
